Move DateExtensions to its own file
