+++ src/Google.Api.CommonProtos/Type/DatePartial.cs
@@ -71,26 +71,4 @@ namespace Google.Type
                 Day = dateTimeOffset.Day
             };
     }
-
-    /// <summary>
-    /// Extension methods built for <see cref="Date"/>.
-    /// </summary>
-    public static class DateExtensions
-    {
-        /// <summary>
-        /// Converts the <see cref="DateTime.Date"/> part of <see cref="DateTime"/> to <see cref="Date"/>.
-        /// </summary>
-        /// <param name="dateTime">The <see cref="DateTime"/> instance being converted.</param>
-        /// <returns>The <see cref="Date"/>.</returns>
-        public static Date ToDate(this DateTime dateTime) =>
-            Date.FromDateTime(dateTime);
-
-        /// <summary>
-        /// Converts the <see cref="DateTimeOffset.Date"/> part of <see cref="DateTimeOffset"/> to <see cref="Date"/>.
-        /// </summary>
-        /// <param name="dateTimeOffset">The <see cref="DateTimeOffset"/> instance being converted.</param>
-        /// <returns>The converted <see cref="Date"/>.</returns>
-        public static Date ToDate(this DateTimeOffset dateTimeOffset) =>
-            Date.FromDateTimeOffset(dateTimeOffset);
-    }
 }
